refactor: flytter faker-funksjoner til __tests__/testdata.ts
DFK-429
